LayoutTests/imported/w3c/web-platform-tests/custom-elements/reactions/customized-builtins/HTMLModElement-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: testReflectAttribute

Harness Error (FAIL), message = ReferenceError: Can't find variable: testReflectAttribute
